HM-3 patterns (PR fix)

+++ src/main/java/ru/astondevs/homework/arychev/patterns/adapter/PaymentProcessor.java
@@ -1,9 +1,11 @@
 package ru.astondevs.homework.arychev.patterns.adapter;
 
 
+import java.math.BigDecimal;
+
 /**
  * Интерфейс для обработки платежей в нашей системе
  */
 public interface PaymentProcessor {
-    void processPayment(String orderId, double amount);
+    void processPayment(String orderId, BigDecimal amount);
 }
